+++ .changesets/fix_buzzard_sunflower_wildlife_cinema.md
@@ -1,0 +1,5 @@
+### Download JWKS out of band ([Issue #2647](https://github.com/apollographql/router/issues/2647))
+
+This moves the JWKS download in the JWT authentication plugin to a separate task that polls them asynchronously, instead of downloading them on demand when a JWT is verified. This should reduce the latency for the first requests received by the router, and increase reliability by removing the tower `Buffer` usage.
+
+By [@Geal](https://github.com/Geal) in https://github.com/apollographql/router/pull/2648
