+++ apollo-router/Cargo.toml
@@ -66,7 +66,6 @@ clap = { version = "4.1.4", default-features = false, features = [
 console-subscriber = { version = "0.1.8", optional = true }
 ci_info = {version="0.14.10", features=["serde-1"] }
 dashmap = { version = "5.4.0", features = ["serde"] }
-deduplicate = "0.3.6"
 derivative = "2.2.0"
 derive_more = { version = "0.99.17", default-features = false, features = [
     "from",
